+++ Changelog.md
@@ -5,6 +5,7 @@
  * Fixed some bugs in ruby-switch. Thanks to John C Banks. Closes GH-47.
  * Code cleanups thanks to Aris Pikeas. Closes GH-43.
  * The build scripts in passenger-customizable now automatically run `apt-get update` when necessary.
+ * Development headers are now included by default in all Ruby images. Closes GH-44.
 
 ## 0.9.11 (release date: 2014-06-25)